refactor: rename Channels.tsx to index.tsx
